changed function type

+++ tasks/multiplication/multiplication.cpp
@@ -1,6 +1,6 @@
 #include "multiplication.h"
 
-long long Multiply(int a, int b) {
-    long long answer = (long long)a * (long long)b;
+int64_t Multiply(int a, int b) {
+    int64_t answer = (int64_t)a * (int64_t)b;
     return answer;
 }
